+++ mail_check/CHANGELOG.md
@@ -1,3 +1,7 @@
 ## [0.0.1] - 28 Jul 2023
 
 * MailCheck Library Initial Build
+
+## [0.0.3] - 9 Aug 2023
+* Enhance suggestion email
+* Support check email validation with regex